requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
black==24.4.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Django==5.0.4
django-jazzmin==3.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
